javacall refactor
