+++ src/main/java/indi/wenyan/content/handler/CommunicateHandler.java
@@ -3,6 +3,7 @@ package indi.wenyan.content.handler;
 import indi.wenyan.content.block.BlockRunner;
 import indi.wenyan.interpreter.structure.WenyanException;
 import indi.wenyan.interpreter.structure.WenyanNativeValue;
+import indi.wenyan.interpreter.utils.JavacallHandlers;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
@@ -24,7 +25,7 @@ public class CommunicateHandler implements JavacallHandler {
 
     @Override
     public WenyanNativeValue handle(WenyanNativeValue[] wenyanValues) throws WenyanException.WenyanThrowException {
-        Object[] args = JavacallHandler.getArgs(wenyanValues, ARG_TYPES);
+        Object[] args = JavacallHandlers.getArgs(wenyanValues, ARG_TYPES);
         args[0] = Math.max(-10, Math.min(10, (int) args[0]));
         args[1] = Math.max(-10, Math.min(10, (int) args[1]));
         args[2] = Math.max(-10, Math.min(10, (int) args[2]));
