+++ src/main/java/indi/wenyan/content/handler/ExplosionHandler.java
@@ -3,6 +3,7 @@ package indi.wenyan.content.handler;
 import indi.wenyan.content.entity.HandRunnerEntity;
 import indi.wenyan.interpreter.structure.WenyanException;
 import indi.wenyan.interpreter.structure.WenyanNativeValue;
+import indi.wenyan.interpreter.utils.JavacallHandlers;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 
@@ -21,7 +22,7 @@ public class ExplosionHandler implements JavacallHandler {
 
     @Override
     public WenyanNativeValue handle(WenyanNativeValue[] wenyan_args) throws WenyanException.WenyanTypeException {
-        Object[] args = JavacallHandler.getArgs(wenyan_args, ARGS_TYPE);
+        Object[] args = JavacallHandlers.getArgs(wenyan_args, ARGS_TYPE);
         args[0] = Math.max(1, Math.min(20, (double) args[0]));
         holder.level().explode(holder, entity.getX(), entity.getY(), entity.getZ(),
                 (float) (double) args[0], Level.ExplosionInteraction.MOB);
