+++ src/main/java/indi/wenyan/content/handler/LocalCallHandler.java
@@ -2,11 +2,12 @@ package indi.wenyan.content.handler;
 
 import indi.wenyan.interpreter.structure.WenyanException;
 import indi.wenyan.interpreter.structure.WenyanNativeValue;
+import indi.wenyan.interpreter.utils.JavacallHandlers;
 
 public class LocalCallHandler implements JavacallHandler {
-    private final WenyanFunction function;
+    private final JavacallHandlers.WenyanFunction function;
 
-    public LocalCallHandler(WenyanFunction function) {
+    public LocalCallHandler(JavacallHandlers.WenyanFunction function) {
         this.function = function;
     }
