+++ src/main/java/indi/wenyan/content/handler/MoveHandler.java
@@ -3,6 +3,7 @@ package indi.wenyan.content.handler;
 import indi.wenyan.content.entity.HandRunnerEntity;
 import indi.wenyan.interpreter.structure.WenyanException;
 import indi.wenyan.interpreter.structure.WenyanNativeValue;
+import indi.wenyan.interpreter.utils.JavacallHandlers;
 import net.minecraft.world.phys.Vec3;
 
 public class MoveHandler implements JavacallHandler {
@@ -17,7 +18,7 @@ public class MoveHandler implements JavacallHandler {
 
     @Override
     public WenyanNativeValue handle(WenyanNativeValue[] args) throws WenyanException.WenyanThrowException {
-        Object[] newArgs = JavacallHandler.getArgs(args, ARGS_TYPE);
+        Object[] newArgs = JavacallHandlers.getArgs(args, ARGS_TYPE);
         newArgs[0] = Math.max(-20, Math.min(20, (double) newArgs[0]));
         newArgs[1] = Math.max(-20, Math.min(20, (double) newArgs[1]));
         newArgs[2] = Math.max(-20, Math.min(20, (double) newArgs[2]));
