+++ src/main/java/indi/wenyan/content/handler/TouchHandler.java
@@ -2,6 +2,7 @@ package indi.wenyan.content.handler;
 
 import indi.wenyan.interpreter.structure.WenyanException;
 import indi.wenyan.interpreter.structure.WenyanNativeValue;
+import indi.wenyan.interpreter.utils.JavacallHandlers;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Explosion;
 import net.minecraft.world.level.Level;
@@ -21,7 +22,7 @@ public class TouchHandler implements JavacallHandler {
 
     @Override
     public WenyanNativeValue handle(WenyanNativeValue[] wenyan_args) throws WenyanException.WenyanThrowException {
-        Object[] args = JavacallHandler.getArgs(wenyan_args, ARGS_TYPE);
+        Object[] args = JavacallHandlers.getArgs(wenyan_args, ARGS_TYPE);
         args[0] = Math.max(-10, Math.min(10, (int) args[0]));
         args[1] = Math.max(-10, Math.min(10, (int) args[1]));
         args[2] = Math.max(-10, Math.min(10, (int) args[2]));
